+++ packages/oklab_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Added `oklabLinearGradient` helper to build `LinearGradient` values
+  from Oklab colors using sampled perceptual interpolation.
+
 ## 0.0.1
 
 * Initial Flutter integration for Oklab: